+++ src/cloud/doc/RELEASE_NOTES.md
@@ -1,3 +1,18 @@
+## Version 1.0.5
+
+### Improvements
+
+- Added max body size configuration to http server
+
+### Submodules
+
+- range-ai-lib @ v1.0.0
+- range-base-lib @ v1.0.1
+- range-build-tools @ v1.0.0
+- range-cloud-lib @ v1.0.2
+
+---
+
 ## Version 1.0.4
 
 ### Bug Fixes